common: Adding ProducerTable class
The ProducerTable class implements SET and DEL API for Redis Table.

SET is implemented via "HSET" command to add the list of fields and
values into Redis DB.
DEL uses Redis "DEL" command.

In addition, the producer-table implement the HSET in a transaction
using "MULTI" and "EXEC" commands.

In addition it adds OP, VALUE and KEY elements into the tracking
list and trigger a SUBSCRIBE to wake up all listeners

Signed-off-by: Elad Raz <eladr@mellanox.com>