[schema]: Change state table name with suffix STATE_TABLE
In the current schema, the state tabel names have conflicts with the
application table names. e.g. PORT_TABLE. It will cause issues when
one Orch class subscribes to two of these tables and uses the table
name to distinguish which table it is.

It is good to make distinctive names even if tables are in different
databases.

Signed-off-by: Shu0T1an ChenG <shuche@microsoft.com>