Ignore/skip when timerfd read returns 0 with errno=0.
timerfd read is expected to return 8 or for any other value appropriate
non-zero errno is expected. But due to a kernel bug, that is found to
occur in S6100 platform, there could be a return value of 0, with errno=0.

Per DELL, this implies timer fired but not forwarded, so take it as
false alarm, and next immediate read is expected to succeed.

Made the fix to meet it.