Fix memory leak in pyext when Selectable is returned to Python (#343)
We have a memory leak in pyext. As result ledd, lddpd, and pmon could use a lot of memory.

I read swig documentation in many places + some source code, but not everything.
As I understood from the documentation SWIG needs to have a mechanism to garbage
collect objects, which are created inside of C++ method, were given outside, and
it is user responsibility to remove the objects. To address such objects with
python swig introduced *_OWN flag. So by documentation and sources I read I think
we definitely don't need *_OWN flag, because we don't want the selectabl object
to be removed, after we receive it from Select.