[logger] Fix Kernel GP seen for any short-lived *syncd process (#444)
Fixes: Azure/sonic-buildimage#6103
This PR is to fix the Kernel GP errors that are seen in any short-lived process within swss.
Issue:
As part of *syncd initialization, Logger::linkToDbNative function is called and a thread is started. When the main *syncd process terminates the destructor Logger::~Logger simply detaches the SettingThread. The exiting main process deletes the static variables. Fault is hit when the detached thread (still executing in the infinite loop) tries to access these freed up variables.
Fix:
Before exiting the main process, set the flag in Logger destructor to signal the detached thread that the main process is finishing up. Instead of detaching the thread (which leaves this thread access the undefined static variables), join the SettingThread thread.