[swig] allow threads (#477)
By default, SWIG does not enable support for multithreaded Python applications. More specifically, the Python wrappers generated by SWIG will not release the Python's interpreter's Global Interpreter Lock (GIL) when wrapped C/C++ code is entered. Hence, while any of the wrapped C/C++ code is executing, the Python interpreter will not be able to run any other threads, even if the wrapped C/C++ code is waiting in a blocking call for something like network or disk IO.

ref: http://www.swig.org/Doc4.0/Python.html#Python_multithreaded
But the solution is for swig 4.0. The PR fix it in swig 3.0.