Use build profiles instead of distro version for Python 2 binding build (#558)
Instead of using the distro version to determine if the Python 2
bindings should be built, use deb profiles instead to specify if they
should be built or not. The main reason of this is that normal .install
and .dirs files can be used for both the Python 2 and Python 3 bindings,
instead of .install/.dir files for Python 3 and debian/rules for Python

More details on build profiles are [here](https://wiki.debian.org/BuildProfileSpec).