Fix minor issues found during static code scan
Fix using uninitialized value when calling Compare on unit tests.
Fix Overrunning array of 16 bytes at byte offset 16 by dereferencing pointer ipa.ip_addr.ipv6_addr + mid + 1 on this case we call memset with a pointer to outside the block. This shouldn't cause any issues due to the fact that we call it with size to be written 0, but still should not be done.