Switch to using stock gcovr 5.2 (#726)
* Switch to using stock gcovr 5.2

The custom branch we were using previously has since been deleted. That
branch appears to have some fix for searching for the source file (for a
gcda file) recursively. I don't know if it's needed or not today, but
using the stock gcovr 5.2 (from the official repo) appears to work.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>