Temporarily bump up time limit for sairedis build
Looks like the C++ code generated by SWIG 4.0 takes much longer to
compile than the C++ code generated by SWIG 3.0 for some reason. For
this reason, bump up the time limit for the sairedis build to 90
minutes.

There's some fixes that can be done on the sairedis side, but that needs
swss-common to be properly building.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>