Update sonic-swss-common pipeline build for Bullseye (#716)
* Don't install python-swsscommon in docker-sonic-vs

This is in preparation for docker-sonic-vs being upgraded to Bullseye,
where Python 2 is no longer available.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Change AzP to build for Bullseye only

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Don't install python-swsscommon_*.deb.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Use pytest-3 instead of pytest

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Clean up build rules to only support Bullseye builds

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Update Microsoft repo URL for Bullseye

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Fix sairedis build for Bullseye

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Fix swss build.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Temporarily bump up time limit for sairedis build

Looks like the C++ code generated by SWIG 4.0 takes much longer to
compile than the C++ code generated by SWIG 3.0 for some reason. For
this reason, bump up the time limit for the sairedis build to 90
minutes.

There's some fixes that can be done on the sairedis side, but that needs
swss-common to be properly building.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Replace deb packages instead of remove and install

In updating docker-sonic-vs, instead of removing the existing packages
first and then installing the current version of packages, just replace
the existing deb packages. This gets around a dependency issue due to
sonic-eventd being in the docker-config-engine-bullseye layer.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Empty commit to make new commit hash.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

---------

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>