Update sonic-swss-common for compilation on Bookworm (#825)
* Use json.hpp from nlohmann-json-dev instead of local version

This header file comes from an external package, and a very old version of the
header file has been checked into this repo. This will cause problems for the
upcoming Bookworm upgrade.

Change references to the header file to use the Debian package
nlohmann-json-dev, and remove json.hpp from this repo.