[azp] Only grab libyang v1 libs for AZP jobs (#1031)
Limit grabbed libyang libraries to only v1.

Upgrading to libyang v3 (sonic-net/sonic-buildimage#21679) is causing the swss-common pipeline build to fail. The pipeline assumes that any package starting with libyang_ or libyang- is a libyang v1 package and tries to install it. However, the dev header package for libyang v3 starts with libyang-, so the pipeline tries to download/install it as well, which causes the build to fail since it depends on the base libyang v3 package which is not grabbed because it starts with libyang3.