Read tables in insertion order in ConfigDBConnector.mod_config()
When the SWSS SDK was ported to C++, one of the effects is that the map of tables passed in to mod_config() is now converted to a std::map by SWIG. Iteration over std::map always happens in lexical order of the table names, which causes issues for frrcfgd when one table depend on the contents of the another (e.g. PREFIX depends on PREFIX_SET, but PREFIX sorts before PREFIX_SET.)

This adds an overload to mod_config() that takes in an unsorted_map of the config data, along with a vector specifying the order in which to read the tables. The Python implementation is adjusted to always use the new function, to preserve the dictionary insertion order.

Fixes https://github.com/sonic-net/sonic-buildimage/issues/22974