Allow using Boost 1.83 for compilation
With the base image upgrade to Trixie, Bookworm-based containers will
need to use Boost 1.83. This is because of an incompatibility between
rsyslog_plugin that uses Boost 1.83 on Trixie and the eventd
container that uses Boost 1.74 on Bookworm; specifically there is an
incompatiblity with serialization of objects between the two versions of
Boost.

Because of this, allow applications to compile with Boost 1.83 headers.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>