Fix trixie compilation (#1040)
Make some changes to fix compilation for Debian Trixie. This includes:

Don't mark m_buffer as const, since the memory that it's pointing to is anyways modified by BinarySerializer.
Add support for compiling with libhiredis 1.1.0.
Add a missing include for <stdexcept>.
Use SWIG_AppendOutput instead of SWIG_Python_AppendOutput, as the latter now takes another parameter to indicate if the function's return type is void, and the recommendation appears to be to just use SWIG_AppendOutput.
Add a workaround for GCC complaining about attributes being ignored in a template argument when passing in pclose as a function pointer into std::unique_ptr.[1]
[1] This is based on https://stackoverflow.com/a/76867913